+++ src/components/footer/footer_complete.jsx
@@ -1,0 +1,14 @@
+import "./footer.css";
+
+function FooterComplete({ setTodos }) {
+  function handleClear() {
+    setTodos((prevTodos) => prevTodos.filter((todo) => !todo.checked));
+  }
+  return (
+    <div className="clear" onClick={handleClear}>
+      Clear completed
+    </div>
+  );
+}
+
+export default FooterComplete;
