+++ src/components/listbox/listbox_delete.jsx
@@ -1,0 +1,10 @@
+import "./listbox.css";
+
+function ListboxDelete({ todo, setTodos }) {
+  const handleDelete = () => {
+    setTodos((prevTodos) => prevTodos.filter((item) => item !== todo));
+  };
+  return <div className="delete" onClick={handleDelete} />;
+}
+
+export default ListboxDelete;
