+++ src/components/textbox/textbox_add.jsx
@@ -1,0 +1,7 @@
+import "./textbox.css";
+
+function TextboxAdd({ setTodos, text, setText, handleSubmit }) {
+  return <div className="input_button" onClick={handleSubmit} />;
+}
+
+export default TextboxAdd;
