+++ src/components/textbox/textbox_check.jsx
@@ -1,0 +1,15 @@
+import "./textbox.css";
+
+function TextboxCheck({ todos, setTodos }) {
+  const checkAll = () => {
+    const allChecked = todos.every((todo) => todo.checked);
+    setTodos(
+      todos.map((todo) => {
+        return { ...todo, checked: !allChecked };
+      })
+    );
+  };
+  return <div className="select_all" onClick={checkAll} />;
+}
+
+export default TextboxCheck;
