+++ src/components/textbox/textbox_input.jsx
@@ -1,0 +1,15 @@
+import "./textbox.css";
+
+function TextboxInput({ text, setText }) {
+  return (
+    <input
+      type="text"
+      value={text}
+      className="input_box"
+      placeholder="What needs to be done?"
+      onChange={(e) => setText(e.target.value)}
+    />
+  );
+}
+
+export default TextboxInput;
